requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.3
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
cryptography==41.0.4
Django==4.1.10
django-appconf==1.0.5
django-crispy-forms==2.0
django-formtools==2.4.1
django-jsonview==2.0.0
django-querystring-tag==1.0.3
django-select2==8.1.2
django-storages==1.13.2
frozenlist==1.4.0
idna==3.4
isodate==0.6.1
multidict==6.0.4
pdfkit==1.0.0
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.8.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
